# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sales Force Management API
  version: 1.0.0
  description: |-
    Field sales backend: attendance, beat plans, site visits, customers, products, orders, reporting, offline sync and administration.

    All endpoints are versioned under `/api/v1/`. Everything except the public configuration endpoints (`/privacy/`, `/terms/`, `/app-version/`, `/app-config/`, `/announcements/`) requires a JWT bearer token from `/api/v1/auth/login/`.
paths:
  /api/v1/admin/announcements/:
    get:
      operationId: admin_announcements_list
      description: |-
        Every announcement, live or not.

        **Query** `is_active` · `priority`

        **Responses** `200` · `201` · `400` · `401` · `403`
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAnnouncementList'
          description: ''
    post:
      operationId: admin_announcements_create
      description: |-
        Every announcement, live or not.

        **Query** `is_active` · `priority`

        **Responses** `200` · `201` · `400` · `401` · `403`
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Announcement'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Announcement'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Announcement'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Announcement'
          description: ''
...
  /api/v1/admin/settings/:
    get:
      operationId: admin_settings_retrieve
      description: |-
        Reads and writes the operational settings.

        Each setting reports an `effect`: `enforced` means this server reads it on
        every relevant request, `advisory` means it is stored and served for the
        client to honour. The difference is published because a switch that stores
        a value and changes nothing is worse than no switch.

        `PUT` takes a partial map — send only what changes.

        **Responses** `200` · `400` · `401` · `403`
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
    put:
      operationId: admin_settings_update
      description: |-
        Reads and writes the operational settings.

        Each setting reports an `effect`: `enforced` means this server reads it on
        every relevant request, `advisory` means it is stored and served for the
        client to honour. The difference is published because a switch that stores
        a value and changes nothing is worse than no switch.

        `PUT` takes a partial map — send only what changes.

        **Responses** `200` · `400` · `401` · `403`
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/announcements/:
    get:
      operationId: announcements_retrieve
      description: |-
        Announcements that are active, started and not yet finished.

        Most urgent first. Not paginated: there are never many, and a start-up
        call should not need a second round trip.

        **Responses** `200`
      summary: Active announcements
      tags:
      - announcements
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/app-config/:
    get:
      operationId: app_config_retrieve
      description: |-
        Everything the app needs to configure itself.

        Values come from the administration settings store, except the ones the
        server genuinely enforces — those are read from where they take effect, so
        this endpoint cannot describe behaviour the server does not have.
        `enforced_by_server` names which are which.

        **Responses** `200`
      summary: Application configuration for the client
      tags:
      - app-config
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/app-version/:
    get:
      operationId: app_version_retrieve
      description: |-
        What the client should be running, and whether this one must update.

        **Query** `platform` (`android`, `ios`, `web`; default android) ·
        `current_version` — send it to get a verdict rather than a description.

        With `current_version`, the response carries `update_status`, one of
        `up_to_date`, `update_available` or `update_required`. The rule that
        decides which is server-side on purpose: a client that computed it could
        not be updated to change it.

        **Responses** `200` · `400` · `404`
      summary: The release the client should be running
      parameters:
      - in: query
        name: current_version
        schema:
          type: string
        description: Send it to receive an update_status verdict.
      - in: query
        name: platform
        schema:
          type: string
        description: android, ios or web
      tags:
      - app-version
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/auth/login/:
    post:
      operationId: auth_login_create
      description: |-
        Sign in with a mobile number, email address or employee code.

        Accepts any of the three as `identifier`; the authentication backend works
        out which one it is. Returns an access/refresh pair together with the full
        user profile, so the client needs one round trip rather than two.

        **Request**  `{"identifier": "9876543210", "password": "…"}`

        **Responses**
        * `200` — `{access, refresh, user, must_change_password}`
        * `400` — a field is missing
        * `401` — no active account matches those credentials
        * `429` — too many attempts (10 per minute per client)
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Login'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
          description: ''
  /api/v1/auth/logout/:
    post:
      operationId: auth_logout_create
      description: |-
        Sign out by blacklisting the caller's refresh token.

        The access token is not revoked — it is stateless and short-lived and
        expires on its own. The refresh token is what keeps a session alive, so
        that is what gets killed.

        **Request**  `{"refresh": "…"}`

        **Responses**
        * `205` — signed out, the client should clear its stored tokens
        * `400` — the token is invalid, already blacklisted, or belongs to someone else
        * `401` — missing or invalid access token
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Logout'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Logout'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Logout'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Logout'
          description: ''
...
  /api/v1/auth/refresh/:
    post:
      operationId: auth_refresh_create
      description: |-
        Exchange a refresh token for a new access token.

        Refresh rotation is on, so the response also carries a **new** refresh
        token and the one just used is blacklisted. Clients must store the new
        refresh token; replaying the old one fails.

        **Request**  `{"refresh": "…"}`

        **Responses**
        * `200` — `{access, refresh}`
        * `401` — the token is expired, malformed or already blacklisted
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
        required: true
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
...
  /api/v1/privacy/:
    get:
      operationId: privacy_retrieve
      description: |-
        The privacy policy in force.

        **Responses**
        * `200` — the current version
        * `404` — nothing published yet
      summary: The privacy policy in force
      tags:
      - privacy
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/terms/:
    get:
      operationId: terms_retrieve
      description: |-
        The terms and conditions in force.

        **Responses** `200` · `404`
      summary: The terms and conditions in force
      tags:
      - terms
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
components:
  schemas:
...
    Announcement:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 150
        message:
          type: string
        priority:
          $ref: '#/components/schemas/PriorityEnum'
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
        is_live:
          type: boolean
          readOnly: true
        created_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - created_by
      - created_by_name
      - id
      - is_live
      - message
      - title
      - updated_at
...
    Login:
      type: object
      description: |-
        Exchanges one identifier plus a password for a token pair.

        The input field is `identifier` rather than `employee_code`: the
        authentication backend accepts a mobile number, an email address or an
        employee code, and the API should not pretend otherwise.
      properties:
        identifier:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
      required:
      - identifier
      - password
    Logout:
      type: object
      description: Validates that a refresh token is usable and belongs to the caller.
      properties:
        refresh:
          type: string
          writeOnly: true
      required:
      - refresh
...
    PaginatedAnnouncementList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Announcement'
...
    PriorityEnum:
      enum:
      - low
      - normal
      - high
      - critical
      type: string
      description: |-
        * `low` - Low
        * `normal` - Normal
        * `high` - High
        * `critical` - Critical
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: /
  description: This server